added description and likes

+++ sub_term.py
@@ -9,14 +9,34 @@ import argparse
 import sys
 
 
+def get_month(m):
+    switch = {
+        1: "Jan",
+        2: "Feb",
+        3: "March",
+        4: "Apr",
+        5: "May",
+        6: "June",
+        7: "July",
+        8: "Aug",
+        9: "Sept",
+        10: "Oct",
+        11: "Nov",
+        12: "Dec",
+    }
+    return switch.get(m)
+
+
 class Vod:
     def __init__(self, entry):
         self.views = entry.media_statistics['views']
         self.title = entry.title
         self.channel = entry.author
         self.videoid = entry.yt_videoid
+        self.upload = f"{get_month(entry.published_parsed.tm_mon)} {entry.published_parsed.tm_mday}, {entry.published_parsed.tm_year}"
+        self.description = entry.summary
+        self.likes = float(entry.media_starrating['average']) * 20
         self.published = entry.published
-        self.thumbnail = entry.media_thumbnail[0]['url']
 
 
 THREADNUM = 12
@@ -111,15 +131,16 @@ if __name__ == "__main__":
         sub_vods.append(z.channel + " "*(21-len(z.channel)) +
                         '- ' + z.title + " "*(56-len(z.title)) + '--- ' + z.videoid)
 
+    # if new videos, clear thumbnail cache
+    if(args.load == False):
+        os.system('rm -rf ' + path + '/cache/*.jpg')
+
     fzf = FzfPrompt()
     chosen = fzf.prompt(
         sub_vods, '--preview "python3 ' + path + '/preview.py {}" --preview-window="left:40%:noborder:wrap"')
     chosen = chosen[0]
     chosen = chosen[-11:]
     print('Opening Player: https://www.youtube.com/watch?v=' + chosen)
-
-    # clear thumbnail cache
-    os.system('rm -rf ' + path + '/cache/*.jpg')
 
     #
     # Play video
